Create and manage crud operations

+++ domain/src/main/java/hector/ruiz/domain/PhotoUi.kt
@@ -3,7 +3,7 @@ package hector.ruiz.domain
 import java.time.OffsetDateTime
 
 data class PhotoUi(
-    var id: Int = 0,
+    var id: Long = 0L,
     val date: OffsetDateTime,
     val path: String
 )
